container_b/requirements.txt
Werkzeug==2.2.3
Flask==2.2.3
Flask-SocketIO==5.3.2
Flask-CORS==3.0.10
eventlet==0.33.3
pika==1.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.3
